+++ src/1.getting_started/7.1.camera_circle/camera_circle.cpp
@@ -269,7 +269,7 @@ int main()
         // 平移：看起来物体后退
         u_View = glm::translate(u_View, glm::vec3(0.0f, 0.0f, viewTranslate));
         // 使用一个简单的透视投影
-        float aspectRatio = SCREEN_WIDTH / SCREEN_HEIGHT;
+        float aspectRatio = (float)SCREEN_WIDTH / (float)SCREEN_HEIGHT;
         u_Projection = glm::perspective(glm::radians(fov), aspectRatio, 0.1f, 1000.0f);
 
 
@@ -327,6 +327,7 @@ void frame_buffer_size_callback(GLFWwindow *window, int width, int height)
 {
     glViewport(0, 0, width, height);
 }
+
 void processInput(GLFWwindow *window)
 {
     bool key_escape_pressed = glfwGetKey(window, GLFW_KEY_ESCAPE) == GLFW_PRESS;
